Extract methods to utils
